requirements.txt
Django==2.2.6
numpy==1.18.1
scikit-surprise==1.0.6
pandas==0.25.3